refactor: convert vlrgg_tracker.py to extensions/vlrgg/ package
Split 1001-line monolith into focused modules:
- _common.py: config, logger, MongoDB helpers, shared constants
- api.py: ApiMixin with VLR.gg fetch and data parsing
- embeds.py: EmbedsMixin with embed builders
- notifications.py: NotificationsMixin with background tasks
- __init__.py: VlrggExtension composed from all mixins

Moves inline `from src.vlrgg import ...` import to module top-level.
